requirements.txt
fastapi
uvicorn
pydantic
streamlit
google-generativeai
requests
beautifulsoup4
